tidying up conversations
