+++ conversations/helpdesk/api.py
@@ -5,6 +5,7 @@ from django.utils import timezone
 
 from conference.models import Conference
 
+from conversations.common_actions import get_actionable_threads
 from conversations.models import Thread, Message
 
 
@@ -41,3 +42,19 @@ def create_new_support_request(conference, requested_by, title, content,
             attachment.save()
 
     return thread, message
+
+
+def get_actionable_helpdesk_threads(conference):
+    """
+    Returns threads that require reply or further work/update from the staff.
+    """
+    return get_actionable_threads(conference).filter(
+        category=Thread.CATEGORIES.HELPDESK
+    )
+
+
+def get_all_helpdesk_threads(conference):
+    return Thread.objects.filter(
+        conference=conference,
+        category=Thread.CATEGORIES.HELPDESK,
+    ).order_by('status')
