+++ templates_jinja/ep19/conversations/_staff_threads_table.html
@@ -1,0 +1,28 @@
+{% macro threads_table(ThreadFilters, threads) -%}
+{# TODO: make this table nice with material design #}
+    <table>
+        <tr>
+            <th>id</th>
+            <th>title</th>
+            <th><a href='?{{ ThreadFilters.order_by_status }}'>status</a></th>
+            <th><a href='?{{ ThreadFilters.order_by_priority }}'>priority</a></th>
+            <th><a href='?{{ ThreadFilters.order_by_last_message }}'>last message</a></th>
+            <th>people</th>
+        </tr>
+
+        {% for thread in threads %}
+        <tr>
+            <td>{{ thread.id }}</td>
+            <td><a href='{{ thread.get_staff_url() }}'>{{ thread.title }}</a></td>
+            <td class='status-bg-{{ thread.status }}'>{{ thread.get_status_display() }}</td>
+            <td class='priority-bg-{{ thread.priority }}'>{{ thread.get_priority_display() }}</td>
+            <td title="{{ thread.last_message_date }}">{{ timesince(thread.last_message_date) }} ago</td>
+            <td>
+                {% for user in thread.people_involved() %}
+                {{ user }},
+                {% endfor %}
+            </td>
+        </tr>
+        {% endfor %}{# thread in threads #}
+    </table>
+{%- endmacro %}
