+++ templates_jinja/ep19/conversations/base.html
@@ -10,12 +10,16 @@ td { padding: .5em; border: 1px solid #eee; }
 th { padding: .5em; text-align: center; background: #eee }
 
 /* TODO - better names? color blind versions? */
-.status-bg-0 { background-color: green; color: #fff; }
-.status-bg-1 { background-color: green; color: yellow; }
-.status-bg-2 { background-color: pink; }
-.status-bg-3 { background-color: orange; }
-.status-bg-4 { background-color: red; }
-.status-bg-5 { background-color: grey; }
+.status-bg-0 { padding: .5em; background-color: green; color: #fff; }
+.status-bg-1 { padding: .5em; background-color: green; color: yellow; }
+.status-bg-2 { padding: .5em; background-color: pink; }
+.status-bg-3 { padding: .5em; background-color: orange; }
+.status-bg-4 { padding: .5em; background-color: red; }
+.status-bg-5 { padding: .5em; background-color: grey; }
+
+.priority-bg-0 { padding: .5em; background-color: green; color: #fff; }
+.priority-bg-10 { padding: .5em; background-color: grey; color: yellow; }
+.priority-bg-100 { padding: .5em; background-color: red; color: yellow;}
 </style>
 
 <div style='padding-top: 100px; max-width: 1024px; margin: 0 auto'>
