+++ templates_jinja/ep19/conversations/helpdesk/inbox.html
@@ -1,5 +1,7 @@
 {% extends "ep19/conversations/base.html" %}
 
+{% import "ep19/conversations/_staff_threads_table.html" as table %}
+
 {% block breadcrumbs %}
 {{ super() }}
 &raquo; <a href='{{ url("staff_helpdesk:all_threads") }}'>Helpdesk</a>
@@ -7,35 +9,10 @@
 {% endblock breadcrumbs %}
 
 {% block conversations_content %}
-    {{ typography.heading("Inbox", 4, class="mb2") }}
+    {{ typography.heading("Helpdesk Inbox", 4, class="mb2") }}
     <p>These are actionable threads – it means either new or replied by the user and waiting for our response/action</p>
 
-    {# TODO make it into a table in material #}
-    <table>
-        <tr>
-            <th>id</th>
-            <th>title</th>
-            <th><a href='?{{ ThreadFilters.order_by_status }}'>status</a></th>
-            <th><a href='?{{ ThreadFilters.order_by_priority }}'>priority</a></th>
-            <th><a href='?{{ ThreadFilters.order_by_last_message }}'>last message</a></th>
-            <th>people</th>
-        </tr>
-
-        {% for thread in actionable %}
-        <tr>
-            <td>{{ thread.id }}</td>
-            <td><a href='{{ thread.get_staff_url() }}'>{{ thread.title }}</a></td>
-            <td class='status-bg-{{ thread.status }}'>{{ thread.get_status_display() }}</td>
-            <td class='priority-bg-{{ thread.priority }}'>{{ thread.get_priority_display() }}</td>
-            <td title="{{ thread.last_message_date }}">{{ timesince(thread.last_message_date) }} ago</td>
-            <td>
-                {% for user in thread.people_involved() %}
-                {{ user }},
-                {% endfor %}
-            </td>
-        </tr>
-        {% endfor %}{# thread in actionble #}
-    </table>
+    {{ table.threads_table(ThreadFilters, actionable) }}
 
     <p>
     Can't see the thread you're looking for?
